pyproject.toml
[project]
name = "hanzi-ids-component-explorer"
version = "1.0.1"
description = "Glyphs plugin for exploring Chinese character IDS components"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "TzuYuan Yin", email = "yintzuyuan@erikyin.net"}
]

dependencies = [
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v"
